refactor(spline): migrate spline-loader to use next.js optimized spline import.
