+++ package.json
@@ -38,6 +38,7 @@
     "@radix-ui/react-toggle-group": "^1.1.2",
     "@radix-ui/react-tooltip": "^1.1.8",
     "@splinetool/react-spline": "^4.0.0",
+    "@splinetool/runtime": "^1.9.82",
     "@stripe/react-stripe-js": "^3.5.1",
     "@stripe/stripe-js": "^6.1.0",
     "bcryptjs": "^3.0.2",
@@ -55,15 +56,16 @@
     "lucide-react": "^0.483.0",
     "mongoose": "^8.13.2",
     "motion": "^12.6.3",
-    "next": "15.2.2",
+    "next": "^15.3.0",
     "next-themes": "^0.4.6",
+    "next-transpile-modules": "^10.0.1",
     "nodemailer": "^6.10.0",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
     "react-resizable-panels": "^2.1.7",
-    "react-syntax-highlighter": "^15.6.1",
+    "react-syntax-highlighter": "^5.8.0",
     "recharts": "^2.15.1",
     "sonner": "^2.0.2",
     "stripe": "^17.7.0",
