+++ temp-spline/package.json
@@ -1,0 +1,66 @@
+{
+  "name": "@splinetool/react-spline",
+  "version": "4.0.0",
+  "packageManager": "yarn@3.3.0",
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "react-native": "./dist/react-spline.js",
+  "types": "./dist/Spline.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/Spline.d.ts",
+      "import": "./dist/react-spline.js"
+    },
+    "./next": {
+      "types": "./dist/next/SplineNext.d.ts",
+      "import": "./dist/react-spline-next.js"
+    }
+  },
+  "scripts": {
+    "start": "vite",
+    "dev": "vite",
+    "serve": "vite preview",
+    "build": "rimraf dist/ && vite build && yarn build-types",
+    "build-types": "tsc --skipLibCheck --isolatedModules false --noEmit false --emitDeclarationOnly --declaration --outDir dist/",
+    "publish": "yarn build && np --no-tests"
+  },
+  "peerDependencies": {
+    "@splinetool/runtime": "*",
+    "next": ">=14.2.0",
+    "react": "*",
+    "react-dom": "*"
+  },
+  "peerDependenciesMeta": {
+    "next": {
+      "optional": true
+    }
+  },
+  "dependencies": {
+    "blurhash": "2.0.5",
+    "lodash.debounce": "4.0.8",
+    "react-merge-refs": "2.1.1",
+    "thumbhash": "0.1.1"
+  },
+  "devDependencies": {
+    "@splinetool/runtime": "^1.3.10",
+    "@types/animejs": "^3.1.12",
+    "@types/lodash.debounce": "^4.0.9",
+    "@types/node": "^20.14.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.0",
+    "animejs": "^3.2.2",
+    "modern-normalize": "^1.1.0",
+    "next": "15.0.0-rc.0",
+    "np": "^10.0.5",
+    "react": "19.0.0-rc-f994737d14-20240522",
+    "react-dom": "19.0.0-rc-f994737d14-20240522",
+    "rimraf": "^5.0.7",
+    "rollup-plugin-rename-node-modules": "^1.3.1",
+    "rollup-preserve-directives": "^1.1.1",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.12"
+  }
+}
